src/gps_module/gps_receiver.py
from dataclasses import dataclass
from typing import List, Optional, Dict, Tuple
import serial
import pynmea2
import time
from datetime import datetime

@dataclass
class SatelliteData:
    prn_code: str                    # Unique Satellite Identifier
    position: Tuple[float, float, float]  # X, Y, Z coordinates
    atomic_timestamp: float          # Nanosecond precision
    transmission_time: float         # Signal send time
    ephemeris_data: Dict[str, float] # Orbital parameters
    almanac_data: Dict[str, any]     # Additional satellite data including:
                                    # - Clock correction
                                    # - Ionospheric data
                                    # - Atmospheric corrections
                                    # - Satellite health
                                    # - Doppler shift

class GPSReceiver:
    def __init__(self, port: str = "/dev/ttyUSB0", baud_rate: int = 9600):
        """Initialize GPS receiver connection."""
        self.port = port
        self.baud_rate = baud_rate
        self.connection = None
        self.satellites: List[SatelliteData] = []
        self.min_satellites = 4
        
    def connect(self) -> bool:
        """Establish connection with GPS module."""
        try:
            self.connection = serial.Serial(self.port, self.baud_rate, timeout=5)
            return True
        except Exception as e:
            print(f"Failed to connect to GPS module: {e}")
            return False
            
    def get_satellite_data(self) -> List[SatelliteData]:
        """Collect raw data from visible satellites."""
        if not self.connection:
            raise ConnectionError("GPS module not connected")
            
        satellites = []
        while len(satellites) < self.min_satellites:
            try:
                raw_data = self.connection.readline().decode('ascii')
                if raw_data.startswith('$GPGSV'):  # Satellites in view
                    msg = pynmea2.parse(raw_data)
                    sat_data = self._process_satellite_message(msg)
                    if sat_data:
                        satellites.append(sat_data)
            except Exception as e:
                print(f"Error reading satellite data: {e}")
                
        self.satellites = satellites
        return satellites
        
    def _process_satellite_message(self, msg) -> Optional[SatelliteData]:
        """Process NMEA message into SatelliteData."""
        try:
            return SatelliteData(
                prn_code=msg.sv_prn_num_1,
                position=self._calculate_satellite_position(msg),
                atomic_timestamp=time.time_ns(),  # Nanosecond precision
                transmission_time=self._get_transmission_time(msg),
                ephemeris_data=self._get_ephemeris_data(msg),
                almanac_data=self._get_almanac_data(msg)
            )
        except Exception as e:
            print(f"Error processing satellite message: {e}")
            return None

    def _calculate_satellite_position(self, msg) -> Tuple[float, float, float]:
        """Calculate precise satellite position using ephemeris data."""
        # Implementation using orbital parameters
        pass

    def _get_transmission_time(self, msg) -> float:
        """Extract precise transmission time from satellite signal."""
        # Implementation for transmission time calculation
        pass

    def _get_ephemeris_data(self, msg) -> Dict[str, float]:
        """Extract orbital parameters from satellite message."""
        return {
            "semi_major_axis": 0.0,
            "eccentricity": 0.0,
            "inclination": 0.0,
            "right_ascension": 0.0,
            "argument_of_perigee": 0.0,
            "mean_anomaly": 0.0
        }

    def _get_almanac_data(self, msg) -> Dict[str, any]:
        """Extract additional satellite data."""
        return {
            "clock_correction": self._get_clock_correction(msg),
            "ionospheric_data": self._get_ionospheric_data(msg),
            "atmospheric_corrections": self._get_atmospheric_corrections(msg),
            "satellite_health": self._get_satellite_health(msg),
            "doppler_shift": self._get_doppler_shift(msg)
        }

    def validate_signal_integrity(self, data: SatelliteData) -> bool:
        """Verify GPS signal hasn't been tampered with."""
        # Implementation for signal validation
        pass

    def calculate_precise_time(self) -> Optional[float]:
        """Calculate precise time using triangulation."""
        if len(self.satellites) < 4:
            return None
            
        # Implementation for time calculation using multiple satellites
        pass 