requirements.txt
pyserial==3.5
pynmea2==1.18.0
pytest==7.3.1 